refactor: Rename `…ws/package/compile/events/api-gateway/lib/apiKeys.js`
