+++ lib/plugins/aws/package/compile/events/api-gateway/index.js
@@ -8,7 +8,7 @@ const memoize = require('memoizee');
 const validate = require('./lib/validate');
 const compileRestApi = require('./lib/restApi');
 const compileRequestValidators = require('./lib/requestValidator');
-const compileApiKeys = require('./lib/apiKeys');
+const compileApiKeys = require('./lib/api-keys');
 const compileUsagePlan = require('./lib/usagePlan');
 const compileUsagePlanKeys = require('./lib/usagePlanKeys');
 const compileResources = require('./lib/resources');
