+++ test/unit/lib/plugins/aws/package/compile/events/api-gateway/lib/api-keys.test.js
